requirements.txt
crewai
crewai-tools
python-dotenv
langchain_groq
langchain_openai